pyproject.toml
[tool.poetry]
name = "misskey-invite-bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "misskey_invite_bot"}]

[tool.poetry.dependencies]
python = "^3.11"
"discord.py" = "^2.3.2"
requests = "^2.32.3"


[tool.poetry.group.development.dependencies]
ruff = "^0.4.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
